pyproject.toml
[tool.poetry]
name = "Privacy-Aware Speech Recognition"
version = "0.0"
description = "Privacy-Aware Speech Recognition"
authors = ["WisamAbbasi <wesam.alabbasi@iit.cnr.it>"]
repository = "https://github.com/sifis-home/flask-whisper-speech-recognition"
license = "MIT"
classifiers = [
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT",
    "Natural Language :: English",
    "Programming Language :: Python :: 3 :: Only",
]
include = ["README.md", "LICENSE.md", "flask-whisper-speech-recognition/data/*"]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
pytest = "^7.2.1"
flask = "2.3.2"
websocket-client = "1.6.1"
rel = "0.4.9"
gTTS = "2.3.0"
requests = "2.28.2"
spacy = "^3.4.0"
numpy = "1.24"
torch = "^1.12.1"
openai-whisper = "20230314"

[tool.poetry.dev-dependencies]
pytest = "^7.2.1"
pre-commit = "^3.0.4"
black = "^23.1.0"
isort = "^5.12.0"
ruff = "^0.0.63"
coverage = {extras = ["toml"], version = "^7.1.0"}

[build-system]
requires = ["poetry-core>=1.0.0", "pip>=22.1.2"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79

[tool.isort]
profile = "black"
line_length = 79

[tool.ruff]
line-length = 79
select = [
    "F401",
    "F403",
]

[tool.coverage.run]
omit = [".*", "*/site-packages/*"]

[tool.coverage.report]
fail_under = 60